+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixed panic on non-existent paths in the manifest.
 - Update MSRV to `1.88`
 - Update Typst to `0.14.0-rc.1`
+- Promote detection of nested tests to an error
 
 ## Fixes
 - Don't panic when trying to update non-persistent tests